requirements.txt
astroid==2.4.1
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
cycler==0.10.0
flake8==3.8.2
Flask==1.1.2
greenlet==0.4.15
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.18.5
opencv-python==4.2.0.34
pickle5==0.0.10
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.2
pynvim==0.4.1
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
